Some minors changes done.
