+++ iris/commons/admin.py
@@ -1,8 +1,7 @@
 from django.contrib import admin
-from .models import DocumentType, PhoneType, Phone, AcademicLevel
+from .models import DocumentType, PhoneType, AcademicLevel
 
 
 admin.site.register(DocumentType)
 admin.site.register(PhoneType)
-admin.site.register(Phone)
 admin.site.register(AcademicLevel)
