+++ iris/members/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 from .models import Disability, Cane, MaritalStatus, PropertyType, Member, \
-Kinship, Kinsman, HousePart, HouseMaterial, Housing, MemberAddress
+    Kinship, Kinsman, HousePart, HouseMaterial, Housing, MemberAddress, Phone
 
 
 admin.site.register(Disability)
@@ -14,3 +14,4 @@ admin.site.register(HousePart)
 admin.site.register(HouseMaterial)
 admin.site.register(Housing)
 admin.site.register(MemberAddress)
+admin.site.register(Phone)
